g3: Use interruptible waits to fix high load average
Based on @cyanogen's commit this replaces uninterruptible waits in
worker threads with their interruptible versions to bring the load
back down to zero at idle.

To make sure that the waits can never be skipped by interrupts, the
return values are checked and jumped back when the event has not yet
finished.

Change-Id: I14bad4af135b1594e235e2252cb484cb34d91054

Conflicts:
	arch/arm/mach-msm/mpm-of.c
	arch/arm/mach-msm/rpm-smd.c
	drivers/cpufreq/cpufreq_ondemand.c
	drivers/slimbus/slim-msm-ngd.c
	drivers/video/msm/mdss/mdss_fb.c

Signed-off-by: Eliminater74 <eliminater74@gmail.com>